% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{crew-package}
\alias{crew-package}
\alias{crew}
\title{crew is a job scheduler for R.}
\description{
crew is a job scheduler for R.
}
\references{
<https://github.com/wlandau/crew>
}
\author{
William Michael Landau \email{will.landau@gmail.com}
}
