% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{example_schedule}
\alias{example_schedule}
\title{Return an example schedule for [hire()].}
\usage{
example_schedule(format = c("tibble", "igraph"))
}
\arguments{
\item{format}{Format of the output. If `"tibble"`, then the output
will be a tibble with columns `from` and `to` to specify dependencies
among the job IDs. For example, in `example_schedule(format = "tibble")`,
the top row tells us that the job named `small` must complete before
the job `regression_small` can begin. For isolated jobs (with no dependencies
and nothing downstream) just include rows where `from` and `to` are the same.
Ultimately, these tibbles
and data frames are converted into `igraph` objects, which are faster
to work with. But the conversion could take time for large schedules.
For the sake of speed, you may want to supply the `igraph`
directly to [hire()].
To get an example igraph, call `example_schedule(format = "igraph")`.}
}
\value{
an example value for the `schedule` argument to [hire()].
}
\description{
Return an example schedule for [hire()].
}
\examples{
# Tibbles and data frames are easy to look at...
example_schedule(format = "tibble")
# But in `hire()`, schedules are ultimately converted into `igraph`
# objects, which could take some time for large schedules.
example_schedule(format = "igraph")
}
\seealso{
[hire()], [example_workload()]
}
