% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{example_workload}
\alias{example_workload}
\title{Return an example workload for [hire()].}
\usage{
example_workload(format = c("tibble", "environment"))
}
\arguments{
\item{format}{Format of the output. If `"tibble"`, then the output
will be a tibble with columns `job` and `command` to specify the jobs
to run. The `job` column has alphanumeric job IDs, and the `command`
column has langauge objects or expressions to evaluate.
Ultimately, these tibbles
and data frames are converted into environments because jobs
are much faster to look up that way.
But the conversion could take time for large workloads.
For the sake of speed, you may want to supply the environment
to [hire()] directly.
To get an example environment,
call `example_schedule(format = "environment")`.}
}
\value{
a named list of language objects (jobs to run)
}
\description{
Return an example workload for [hire()].
}
\examples{
# Tibbles and data frames are easy to look at...
example_workload(format = "tibble")
# But in `hire()`, they are ultimately converted into environments
# because objects are much faster to look up that way.
# For
envir <- example_workload(format = "environment")
envir
ls(envir)
envir$small
}
\seealso{
[hire()], [example_schedule()]
}
