% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master.R
\name{run_master}
\alias{run_master}
\title{Internal function to run the master process}
\usage{
run_master(cache, schedule)
}
\arguments{
\item{cache}{a `storr` cache to communicate with the workers}

\item{schedule}{an `igraph` of job dependencies}
}
\value{
nothing
}
\description{
Do not call this function directly.
  It is only exported to make available to `callr::r_bg()`.
  It is not an interface function. The API is unstable.
}
\keyword{internal}
