% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduler.R
\name{schedule}
\alias{schedule}
\title{Run jobs in topological order}
\usage{
schedule(graph)
}
\arguments{
\item{graph}{An `igraph` object with `names` and `code` attributes.}
}
\description{
Run jobs in topological order
}
