% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{workers-package}
\alias{workers-package}
\alias{workers}
\title{workers is a job scheduler for R.}
\description{
workers is a job scheduler for R.
}
\references{
<https://github.com/wlandau/workers>
}
\author{
William Michael Landau \email{will.landau@gmail.com}
}
