% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{crew-package}
\alias{crew-package}
\alias{crew}
\title{crew is a job scheduler for R.}
\description{
crew is a job scheduler for R.
}
\examples{
\dontrun{
withr::with_dir(tempfile(), {
# Run all the jobs in the workload in the
# correct order given in the schedule.
# Use 2 persistent workers.
hire(
  workload = workload = example_workload(),
  schedule = schedule = example_schedule(),
  fun = parallel::mclapply,
  workers = 2,
  mc.cores = 2
)
})
}
}
\references{
<https://github.com/wlandau/crew>
}
\author{
William Michael Landau \email{will.landau@gmail.com}
}
