% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{example_schedule}
\alias{example_schedule}
\title{Return an example schedule for [hire()].}
\usage{
example_schedule()
}
\value{
an `igraph` object of job IDs corresponding to
  `names(example_workload())`.
}
\description{
Return an example schedule for [hire()].
}
\examples{
example_schedule()
}
\seealso{
[hire()], [example_workload()]
}
