% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{example_workload}
\alias{example_workload}
\title{Return an example workload for [hire()].}
\usage{
example_workload()
}
\value{
a named list of language objects (jobs to run)
}
\description{
Return an example workload for [hire()].
}
\examples{
example_workload()
}
\seealso{
[hire()], [example_schedule()]
}
