% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hire.R
\name{hire}
\alias{hire}
\title{Hire a crew of R workers.}
\usage{
hire(workload, schedule, workers = 1, type = "lapply", fun = lapply, ...)
}
\arguments{
\item{workload}{named list of jobs.
The names are job IDs and the values are expressions or language objects.}

\item{schedule}{an `igraph` object linking the job IDs together.
This is how `crew` knows how to execute some jobs before other jobs.}

\item{type}{character scalar naming the type of crew to hire.
Currently, `"lapply"` is the only crew type.}

\item{fun}{an `lapply`-like function to use for persistent workers
if `type` is `"lapply"`.}

\item{...}{additional arguments.}
}
\value{
nothing
}
\description{
Hire a crew of R workers.
}
\examples{
\dontrun{
withr::with_dir(tempfile(), {
attach(example_args()) # Get an example workload and schedule.
# Run all the jobs in the workload in the
# correct order given in the schedule.
# Use 2 persistent workers.
hire(
  workload = workload,
  schedule = schedule,
  fun = parallel::mclapply,
  workers = 2,
  mc.cores = 2
)
})
}
}
